Extrais <ActionsComplementaires />
… car dans le commit suivant, on va l'augmenter, donc on se fait de la place.
